Tabular: Moved config files and predictor
