+++ docs/tutorials/tabular_prediction/tabular-multimodal-text-others.md
@@ -222,7 +222,7 @@ However, we should notice that the training time will be longer and the inferenc
 
 ```{.python .input}
 from autogluon.text.text_prediction.text_prediction import ag_text_prediction_params
-from autogluon.tabular.task.tabular_prediction.hyperparameter_configs import get_hyperparameter_config
+from autogluon.tabular.configs.hyperparameter_configs import get_hyperparameter_config
 import copy
 
 text_nn_params = ag_text_prediction_params.create('default_electra_base_no_hpo')
