+++ tabular/src/autogluon/tabular/__init__.py
@@ -1,6 +1,7 @@
 import logging
 
 from .features import FeatureMetadata
-from .task.tabular_prediction import TabularDataset, TabularPredictor, TabularPrediction
+from .task.tabular_prediction import TabularDataset, TabularPrediction
+from .predictor import TabularPredictor
 
 logging.basicConfig(format='%(message)s')  # just print message in logs
