+++ tabular/src/autogluon/tabular/predictor/__init__.py
@@ -1,0 +1,1 @@
+from .predictor import TabularPredictor
