+++ tabular/src/autogluon/tabular/task/tabular_prediction/__init__.py
@@ -1,3 +1,2 @@
 from .tabular_prediction import TabularPrediction
 from .dataset import TabularDataset
-from .predictor import TabularPredictor
