+++ tabular/src/autogluon/tabular/task/tabular_prediction/predictor_legacy.py
@@ -10,7 +10,7 @@ import pandas as pd
 import networkx as nx
 
 from .dataset import TabularDataset
-from .hyperparameter_configs import get_hyperparameter_config
+from ...configs.hyperparameter_configs import get_hyperparameter_config
 from autogluon.core.utils import plot_performance_vs_trials, plot_summary_of_models, plot_tabular_models, verbosity2loglevel
 from ...utils import BINARY, MULTICLASS, REGRESSION, get_pred_from_proba
 from ...learner import AbstractLearner as Learner  # TODO: Keep track of true type of learner for loading
